Added all files related to find.c
BubbleSort.c is testing a bubble sort algorithm to be used in the helpers.c file. The generate.c file creates random numbers, which can be piped into find.c through command line arguments. Find.c will search for a number in a provided list, and report to the user whether or not the number was found.